+++ R/fill-na.R
@@ -8,7 +8,8 @@
 #' @return The object with missing values filled.
 #' @export
 #' @examples
-#' fill_na(nlist(x = c(2, NA), y = matrix(c(1:3, NA), nrow =)))
+#' fill_na(nlist(x = c(2, NA), y = matrix(c(1:3, NA), nrow = 2)))
+#' fill_na(nlists(nlist(x = c(2, NA)), nlist(x = c(NA_real_, NA))))
 fill_na <- function(x, ...) UseMethod("fill_na")
 
 #' @describeIn fill_na Fill missing values of logical object
@@ -58,5 +59,14 @@ fill_na.nlist <- function(x, value = 0L, ...) {
   chk_unused(...)
   x <- lapply(x, fill_na, value = value)
   class(x) <- "nlist"
+  x
+}
+
+#' @describeIn fill_na Fill missing values of nlist object
+#' @export
+fill_na.nlists <- function(x, value = 0L, ...) {
+  chk_unused(...)
+  x <- lapply(x, fill_na, value = value)
+  class(x) <- "nlists"
   x
 }