+++ NEWS.md
@@ -1,4 +1,4 @@
-- Added `fill_na()` for natomic and nlist objects.
+- Added `fill_na()` for natomic, nlist and nlists objects.
 - Added `as.nlist.numeric()` to convert flattened nlist objects back into original objects.
 - Added `unlist.nlist()` to flatten nlist objects into term named numeric vectors.